separate OCW tests into its own file
